CI: add write permission for packages
When we create a PR from a branch in toverainc/willow, CI can build and
publish the Willow build container fine. However, when opening a PR from
a fork, CI fails when trying to publish the build container:

------
 > pushing ghcr.io/toverainc/willow-test:pr-1 with docker:
------
ERROR: denied: installation not allowed to Write organization package
Error: buildx failed with: ERROR: denied: installation not allowed to Write organization package

Add write permissions for packages to the build_container job to fix
this.